refactor(svelte): move geoms/, assembly/, chrome/, a11y/ into feature directories
