+++ packages/svelte/src/lib/a11y/CanvasA11y.svelte
