+++ packages/svelte/src/lib/a11y/canvas-a11y.ts
