+++ packages/svelte/src/lib/chrome/ToolRail.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { InteractionTool, ZoomDomains } from "./interaction.js";
+  import type { InteractionTool, ZoomDomains } from "../interaction.js";
 
   type RecoveryInputSource = "keyboard" | "pointer" | "touch";
