+++ packages/svelte/src/lib/geoms/GeomArea.svelte
@@ -6,7 +6,7 @@
    */
   import type { AesInput, AreaParams, StackablePosition } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends AreaParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
