+++ packages/svelte/src/lib/geoms/GeomBar.svelte
@@ -1,12 +1,13 @@
 <script lang="ts">
   /**
-   * <GeomCol> — declaration-only column-layer sugar for <GGPlot> (bars from
-   * pre-computed heights; default position "stack").
-   * Same contract and z-order constraint as <GeomPoint> (decision 0001).
+   * <GeomBar> — declaration-only counting-bar sugar for <GGPlot> (count stat:
+   * do NOT map aes.y; map aes.weight to sum weights; default position
+   * "stack"). Same contract and z-order constraint as <GeomPoint>
+   * (decision 0001).
    */
   import type { AesInput, BarParams, StackablePosition } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends BarParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
@@ -16,5 +17,5 @@
   }
 
   const props: Props = $props();
-  createGeomLayer("col", () => props, ["alpha", "width"]);
+  createGeomLayer("bar", () => props, ["alpha", "width"]);
 </script>
