+++ packages/svelte/src/lib/geoms/GeomCol.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
   /**
-   * <GeomBar> — declaration-only counting-bar sugar for <GGPlot> (count stat:
-   * do NOT map aes.y; map aes.weight to sum weights; default position
-   * "stack"). Same contract and z-order constraint as <GeomPoint>
-   * (decision 0001).
+   * <GeomCol> — declaration-only column-layer sugar for <GGPlot> (bars from
+   * pre-computed heights; default position "stack").
+   * Same contract and z-order constraint as <GeomPoint> (decision 0001).
    */
   import type { AesInput, BarParams, StackablePosition } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends BarParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
@@ -17,5 +16,5 @@
   }
 
   const props: Props = $props();
-  createGeomLayer("bar", () => props, ["alpha", "width"]);
+  createGeomLayer("col", () => props, ["alpha", "width"]);
 </script>
