+++ packages/svelte/src/lib/geoms/GeomDensity.svelte
@@ -7,7 +7,7 @@
    */
   import type { AesInput, DensityParams } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends DensityParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
