+++ packages/svelte/src/lib/geoms/GeomHistogram.svelte
@@ -7,7 +7,7 @@
    */
   import type { AesInput, BarParams, StackablePosition } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends BarParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
