+++ packages/svelte/src/lib/geoms/GeomPoint.svelte
@@ -19,7 +19,7 @@
     PositionParams,
   } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends PointParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
