+++ packages/svelte/src/lib/geoms/GeomSmooth.svelte
@@ -7,7 +7,7 @@
    */
   import type { AesInput, SmoothParams } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends SmoothParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
