+++ packages/svelte/src/lib/geoms/GeomText.svelte
@@ -6,7 +6,7 @@
    */
   import type { AesInput, PositionParams, TextParams } from "@ggsvelte/spec";
 
-  import { createGeomLayer } from "./geom-factory.svelte.js";
+  import { createGeomLayer } from "./factory.svelte.js";
 
   interface Props extends TextParams {
     /** Layer-level aes (bare-string shorthand allowed); merges over plot aes. */
