+++ packages/svelte/src/lib/geoms/registry.svelte.ts
