+++ packages/svelte/src/lib/plot-zoom-state.svelte.ts
@@ -18,7 +18,7 @@ import type {
   ZoomEvent,
 } from "./interaction.js";
 import { frozenZoomDomains, type ContinuousZoomDomains } from "./plot-geometry.js";
-import { zoomAnnouncement } from "./plot-labels.js";
+import { zoomAnnouncement } from "./assembly/labels.js";
 import {
   applyZoomToSpec,
   buildZoomEvent,
